backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pydantic==2.9.2
python-dotenv
pydantic[email]
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.5.0
python-multipart==0.0.9
flask==3.0.0
flask-cors==4.0.0 